chore : 디렉토리 이름 통일 (delete under bar)
